Adjust /apps/ to point at quasi-localhost
Sandstorm provides a local.sandstorm.io domain so that there is
a domain that provides TXT records for people running the Sandstorm
code on their own machine in a sort of development capacity.

Meanwhile, the /apps/ list contains (before this patch) the plain
localhost name, which if clicked-on, will lead to the message:

```Error looking up DNS TXT records for host localhost: queryTxt ENOTFOUND
```

in the browser.

This patch uses local.sandstorm.io instead.