Add documentation links to nav and footer.
I collapsed "Sign up" and "Download" back into a single "Get Sandstorm" both to make space and because user testing (and personal experience) suggested that having "Download" link into the middle of a page seemed weird. The words "Hosting" and "Download" now appear as links in the footer, for ctrl-f-ability.

I re-arranged the footer items to add a "Documentation" section and moved "Sandstorm Blog" to be under "Community" to make space, but I'm open to other ideas.