Add experiments framework
We inline the script needed to support this in the page <head> so we
can generate stylesheets according to experiment membership *before*
any page content is displayed.  Otherwise, we'd potentially wind up
with flash of unstyled content or page reflow, which is undesirable.

As a result of the above inlining of all the early script, stats.js is no
longer used directly by this site, but remains as a publicly-consumable
endpoint for use by other sites that src it.  This change also exposes
the experiment set (including membership), which should allow us to
maintain a single copy of experiment rules/membership sets that will
be used by all of our properties.

This change does not actually enable any experiments; they will come
in future changes.