Remove apparent debug code that suddenly started producing output.
I'm not sure why this code was here. It looks like I was trying to debug something when implementing pagination. Probably `count` was a variable I was setting elsewhere, but then I stopped setting it. As long as `count` was unset, this code didn't produce any output. But now, with a Jekyll update, `count` is apparently set to `8`, so the number `8` suddenly started appearing all over the blog. Oops.